selinux check

+++ podman-export-cmd.sh
@@ -1,4 +1,10 @@
 cd aap-config-export
+#doesnt work if selinux is enforcing, not sure why, disable for now
+if [ $(getenforce) = "Enforcing" ]
+then
+	echo you must run this with selinux disabled sorry, type "sudo setenforce 0" and retry
+	exit 1
+fi
 podman run -it --rm --name aap-porter -v "$PWD":/mnt localhost/aap-porter:latest bash -c 'cd /mnt/ && ./export.cmdline.sh'
 if [ -d ../flatexport ]
 then
